+++ src/ansys/sherlock/core/lifecycle.py
@@ -712,7 +712,7 @@ class Lifecycle(GrpcStub):
             Name of the life cycle phase to add the thermal event to.
         event_name : str
             Name of the thermal event.
-        num_of_cycles : double
+        num_of_cycles : float
             Number of cycles for the thermal event.
         cycle_type : str
             Cycle type. Options are ``"COUNT"``, ``"DUTY_CYCLE"``, ``"PER_YEAR"``,
@@ -836,9 +836,9 @@ class Lifecycle(GrpcStub):
                     Name of the thermal step.
                 - type : str
                     Type of the thermal step. Options are ``"HOLD"`` and ``"RAMP"``.
-                - time : double
+                - time : float
                     Duration of the thermal step expressed in time units.
-                - temperature : double
+                - temperature : float
                     Temperature of the step expressed in temperature units.
 
 
@@ -991,17 +991,17 @@ class Lifecycle(GrpcStub):
             Name of the life cycle phase to add the harmonic event to.
         event_name : str
             Name of the harmonic event.
-        duration : double
+        duration : float
             Event duration length.
         duration_units : str
             Event duration units. Options are ``"ms"``, ``"sec"``, ``"min"``,
             ``"hr"``, ``"day"``, and ``"year"``.
-        num_of_cycles : double
+        num_of_cycles : float
             Number of cycles for the harmonic event.
         cycle_type : str
             Cycle type. Options are ``"COUNT"``, ``"DUTY_CYCLE"``, ``"PER_YEAR"``,
             ``"PER_DAY"``, ``"PER_HOUR"``, ``"PER_MIN"``, and ``"PER_SEC"``.
-        sweep_rate : double
+        sweep_rate : float
             Sweep rate for the harmonic event.
         orientation : str
             PCB orientation in the format of ``"azimuth, elevation"``. For example,
@@ -1151,9 +1151,9 @@ class Lifecycle(GrpcStub):
             - harmonic_profile_entries : list
                 List of harmonic profile entries consisting of these properties:
 
-                - frequency : double
+                - frequency : float
                     Frequency of the harmonic profile expressed in frequency units.
-                - load : double
+                - load : float
                     Load of the harmonic profile expressed in load units.
                 - triaxial_axis : str
                     Axis that this profile should be assigned to if the harmonic
@@ -1318,12 +1318,12 @@ class Lifecycle(GrpcStub):
             Name of the life cycle phase to add this shock event to.
         event_name : str
             Name of the shock event.
-        duration : double
+        duration : float
             Event duration length.
         duration_units : str
             Event duration units. Options are ``"ms"``, ``"sec"``, ``"min"``, ``"hr"``,
             ``"day"``, and ``"year"``.
-        num_of_cycles : double
+        num_of_cycles : float
             Number of cycles for the shock event.
         cycle_type : str
             Cycle type. Options are ``"COUNT"``, ``"DUTY CYCLE"``,
@@ -1446,12 +1446,12 @@ class Lifecycle(GrpcStub):
                 Name of the shock event.
             - profile_name : str
                 Name of the shock profile.
-            - duration : double
+            - duration : float
                 Pulse duration.
             - duration_units : str
                 Pulse duration units. Options are ``"ms"``, ``"sec"``, ``"min"``, ``"hr"``,
                 ``"day"``, and ``"year"``.
-            - sample_rate : double
+            - sample_rate : float
                 Sample rate.
             - sample_rate_units : str
                 Sample rate units. Options are ``"ms"``, ``"sec"``, ``"min"``, ``"hr"``,
@@ -1468,11 +1468,11 @@ class Lifecycle(GrpcStub):
                     Shape of the shock profile entry. Options are ``"FullSine"``,
                     ``"HalfSine"``, ``"Haversine"``, ``"Triangle"``, ``"Sawtooth"``,
                     ``"FullSquare"``, and ``"HalfSquare"``.
-                - load : double
+                - load : float
                     Load of the profile entry expressed in load units.
-                - freq : double
+                - freq : float
                     Frequency of the profile entry expressed in frequency units.
-                - decay : double
+                - decay : float
                     Decay value of the profile entry.
 
         Examples
