+++ src/ansys/sherlock/core/stackup.py
@@ -172,7 +172,7 @@ class Stackup(GrpcStub):
             Name of the Sherlock project.
         cca_name : str
             Name of the CCA.
-        board_thickness : double
+        board_thickness : float
             Board thickness.
         board_thickness_unit : str
             Units for the board thickness.
@@ -184,17 +184,17 @@ class Stackup(GrpcStub):
             Material for the PCB.
         conductor_layers_cnt : int32
             Number of conductor layers.
-        signal_layer_thickness : double
+        signal_layer_thickness : float
             Signal layer thickness.
         signal_layer_thickness_unit : str
             Units for the signal layer thickness.
-        min_laminate_thickness : double
+        min_laminate_thickness : float
             Minimum thickness of laminate layers.
         min_laminate_thickness_unit : str
             Units for the minimum thickness of laminate layers.
         maintain_symmetry : bool
             Whether to maintain symmetry.
-        power_layer_thickness : double
+        power_layer_thickness : float
             Power layer thickness.
         power_layer_thickness_unit : str
             Units for the power layer thickness.
@@ -317,7 +317,7 @@ class Stackup(GrpcStub):
             ``"SIGNAL"``, ``"POWER"``, or ``"SUBSTRATE"``.
         material : str, optional
             Conductor material. The default is ``""``.
-        thickness : double, optional
+        thickness : float, optional
             Conductor layer thickness. The default is ``0``.
         thickness_unit : str, optional
             Units for the conductor layer thickness. The
@@ -456,7 +456,7 @@ class Stackup(GrpcStub):
             Material grade. The default is ``""``.
         material : str, optional
             Material name. The default is ``""``.
-        thickness : double, optional
+        thickness : float, optional
             Laminate thickness. The default is ``0``.
         thickness_unit : str, optional
             Units for the laminate thickness. The default is ``""``.
@@ -468,9 +468,9 @@ class Stackup(GrpcStub):
 
            - style : str
                Style of the glass construction.
-           - resinPercentage : double
+           - resinPercentage : float
                Resin percentage.
-           - thickness: double
+           - thickness: float
                Thickness.
            - thicknessUnit: str
                Units for the thickness.
